requirements.txt
discord.py==2.7.0
selenium==4.9.1
webdriver-manager==3.8.6
PyNaCl==1.5.0